支持微信布局子元素的空隙添加

+++ app/src/main/java/com/szysky/customize/simageview/util/UIUtils.java
@@ -20,7 +20,7 @@ public class UIUtils {
 
 
 	/** dip转换px */
-	public static int dip2px(Context context, int dip) {
+	public static int dip2px(Context context, float dip) {
 		final float scale = context.getResources().getDisplayMetrics().density;
 		return (int) (dip * scale + 0.5f);
 	}
